Fix leaderboard API UI separation and add live player lives display
